runtime-sdk-macros: Refactor sdk_derive for multiple traits
